chore: ChannelList/utils.ts
